# specs/openapi.yaml
openapi: 3.1.0
info:
  title: AI Customer Service Demo API
  version: 1.0.0
  description: >
    Frozen V1 API draft for the Web-first e-commerce AI customer service demo.
    All endpoints are scoped to an anonymous Demo Workspace.
servers:
  - url: /api
security:
  - DemoWorkspaceToken: []
...
paths:
...
  /shops/{shopId}:
    parameters:
      - $ref: "#/components/parameters/ShopId"
    get:
      tags: [Shops]
      summary: Get a shop
      responses:
        "200":
          description: Shop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
    patch:
      tags: [Shops]
      summary: Update mutable shop fields
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name:
                  type: string
                enabled:
                  type: boolean
      responses:
        "200":
          description: Updated shop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
...
  /shops/{shopId}/products:
    parameters:
      - $ref: "#/components/parameters/ShopId"
    get:
      tags: [Products]
      summary: List products for shop
      parameters:
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /shops/{shopId}/products/sync:
    parameters:
      - $ref: "#/components/parameters/ShopId"
    post:
      tags: [Products]
      summary: Sync dynamic product facts from MockDouyinAdapter
      responses:
        "202":
          description: Product sync accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductSyncResult"
...
  /shops/{shopId}/product-learning-jobs:
    parameters:
      - $ref: "#/components/parameters/ShopId"
    post:
      tags: [Products]
      summary: Start or retry a shop-scoped product learning job
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductLearningInput"
      responses:
        "202":
          description: Product learning job accepted or current job snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductLearningJob"
    get:
      tags: [Products]
      summary: List product learning jobs for a shop
      responses:
        "200":
          description: Product learning job snapshots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductLearningJob"
  /products/{productId}/learn:
    parameters:
      - $ref: "#/components/parameters/ProductId"
    post:
      tags: [Products]
      summary: Force product learning through its shop-scoped job
      parameters:
        - in: query
          name: shopId
          schema:
            type: string
      responses:
        "202":
          description: Product learning job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductLearningJob"
...
  /knowledge:
    get:
      tags: [Knowledge]
      summary: List active and historical knowledge
      parameters:
        - in: query
          name: shopId
          schema:
            type: string
        - in: query
          name: productId
          schema:
            type: string
        - in: query
          name: scope
          schema:
            type: string
            enum: [STORE, PRODUCT]
        - in: query
          name: businessStatus
          schema:
            $ref: "#/components/schemas/KnowledgeBusinessStatus"
      responses:
        "200":
          description: Knowledge items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KnowledgeItem"
    post:
      tags: [Knowledge]
      summary: Create manual knowledge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeCreateInput"
      responses:
        "201":
          description: Knowledge created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeItem"
  /knowledge/{knowledgeId}:
    parameters:
      - $ref: "#/components/parameters/KnowledgeId"
    patch:
      tags: [Knowledge]
      summary: Create a new version of knowledge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeUpdateInput"
      responses:
        "202":
          description: New version indexing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeRevisionResult"
    delete:
      tags: [Knowledge]
      summary: Soft delete knowledge
      parameters:
        - in: query
          name: shopId
          schema:
            type: string
      responses:
        "204":
          description: Soft deleted
  /knowledge/{knowledgeId}/reindex:
    parameters:
      - $ref: "#/components/parameters/KnowledgeId"
    post:
      tags: [Knowledge]
      summary: Retry knowledge indexing
      parameters:
        - in: query
          name: shopId
          schema:
            type: string
      responses:
        "202":
          description: Reindex accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeVersion"
  /knowledge/imports:
    post:
      tags: [Knowledge]
      summary: Upload Excel or CSV knowledge file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [shopId, file]
              properties:
                shopId:
                  type: string
                file:
                  type: string
                  format: binary
      responses:
        "202":
          description: Import job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeImportAccepted"
  /knowledge/imports/{jobId}:
    parameters:
      - $ref: "#/components/parameters/JobId"
    get:
      tags: [Knowledge]
      summary: Get import preview and status
      responses:
        "200":
          description: Import job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeImportJob"
  /knowledge/imports/{importId}/commit:
    parameters:
      - $ref: "#/components/parameters/ImportId"
    post:
      tags: [Knowledge]
      summary: Commit the valid rows of a previewed knowledge import
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                shopId: { type: string }
      responses:
        "201":
          description: Import committed with row outcomes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeImportJob"
  /knowledge/search:
    post:
      tags: [Knowledge]
      summary: Search scoped knowledge with Hybrid RAG
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeSearchInput"
      responses:
        "200":
          description: Retrieval status and evidence snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeSearchResult"
  /knowledge/candidates:
    get:
      tags: [Knowledge]
      summary: List knowledge candidates
      parameters:
        - in: query
          name: shopId
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Candidates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KnowledgeCandidate"
  /knowledge/candidates/{candidateId}/approve:
    parameters:
      - $ref: "#/components/parameters/CandidateId"
    post:
      tags: [Knowledge]
      summary: Approve candidate and create knowledge
      parameters:
        - in: query
          name: shopId
          schema:
            type: string
      responses:
        "202":
          description: Candidate approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationAccepted"
  /knowledge/candidates/{candidateId}/reject:
    parameters:
      - $ref: "#/components/parameters/CandidateId"
    post:
      tags: [Knowledge]
      summary: Reject candidate
      parameters:
        - in: query
          name: shopId
          schema:
            type: string
      responses:
        "204":
          description: Rejected
  /knowledge/conflicts:
    get:
      tags: [Knowledge]
      summary: List knowledge conflicts
      parameters:
        - in: query
          name: shopId
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Conflicts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KnowledgeConflict"
  /knowledge/conflicts/{conflictId}:
    parameters:
      - in: path
        name: conflictId
        required: true
        schema:
          type: string
    get:
      tags: [Knowledge]
      summary: Get one scoped knowledge conflict with immutable version snapshots
      responses:
        "200":
          description: Conflict detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeConflict"
  /knowledge/conflicts/{conflictId}/resolve:
    parameters:
      - in: path
        name: conflictId
        required: true
        schema:
          type: string
    post:
      tags: [Knowledge]
      summary: Resolve a conflict
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resolution]
              properties:
                shopId:
                  type: string
                resolution:
                  type: string
                  enum: [KEEP_LEFT, KEEP_RIGHT, MERGE, CUSTOM]
                customQuestion:
                  type: string
                customAnswer:
                  type: string
      responses:
        "202":
          description: Resolution accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationAccepted"
...
components:
  securitySchemes:
    DemoWorkspaceToken:
      type: apiKey
      in: header
      name: X-Demo-Workspace-Token
...
  parameters:
    ShopId:
      in: path
      name: shopId
      required: true
      schema: { type: string }
    ProductId:
      in: path
      name: productId
      required: true
      schema: { type: string }
...
    KnowledgeId:
      in: path
      name: knowledgeId
      required: true
      schema: { type: string }
    CandidateId:
      in: path
      name: candidateId
      required: true
      schema: { type: string }
    JobId:
      in: path
      name: jobId
      required: true
      schema: { type: string }
    ImportId:
      in: path
      name: importId
      required: true
      schema: { type: string }
...
  schemas:
...
    ShopAIMode:
      type: string
      enum: [AUTO_ALLOWED, ASSIST_ONLY, MANUAL_ONLY]
...
    Shop:
      type: object
      required: [id, name, platform, aiMode, connectionState]
      properties:
        id: { type: string }
        name: { type: string }
        platform: { type: string, enum: [DOUYIN_DEMO] }
        aiMode:
          $ref: "#/components/schemas/ShopAIMode"
        connectionState:
          type: string
          enum: [CONNECTED, RECONNECTING, RECONCILING, DEGRADED, DISCONNECTED]
        enabled: { type: boolean }
        todayConversationCount: { type: integer }
...
    Product:
      type: object
      required: [id, shopId, externalProductId, title, status]
      properties:
        id: { type: string }
        shopId: { type: string }
        externalProductId: { type: string }
        title: { type: string }
        description: { type: string }
        status: { type: string }
        recommendable: { type: boolean }
        learningStatus: { type: string }
        skus:
          type: array
          items:
            $ref: "#/components/schemas/ProductSku"
    ProductSku:
      type: object
      required: [id, externalSkuId, price, inventory]
      properties:
        id: { type: string }
        externalSkuId: { type: string }
        attributes:
          type: object
          additionalProperties:
            type: string
        price: { type: number }
        inventory: { type: integer }
        status: { type: string }
...
    ProductSyncResult:
      type: object
      required: [status, synthetic, productsSynced]
      properties:
        status: { type: string, enum: [SUCCEEDED] }
        synthetic: { type: boolean }
        productsSynced: { type: integer, minimum: 0 }
    ProductLearningInput:
      type: object
      additionalProperties: false
      properties:
        productIds:
          type: array
          items: { type: string }
        retryFailed: { type: boolean }
    ProductLearningJobItem:
      type: object
      required: [productId, status]
      properties:
        productId: { type: string }
        status: { type: string, enum: [PENDING, PROCESSING, SUCCEEDED, FAILED, OUTDATED] }
        reason: { type: string, nullable: true }
    ProductLearningJobTotals:
      type: object
      required: [total, created, updated, skipped, failed]
      properties:
        total: { type: integer, minimum: 0 }
        created: { type: integer, minimum: 0 }
        updated: { type: integer, minimum: 0 }
        skipped: { type: integer, minimum: 0 }
        failed: { type: integer, minimum: 0 }
    ProductLearningJob:
      type: object
      required: [id, shopId, status, totals, items]
      properties:
        id: { type: string }
        shopId: { type: string }
        status: { type: string, enum: [PENDING, RUNNING, PARTIAL_SUCCESS, SUCCEEDED, FAILED] }
        totals:
          $ref: "#/components/schemas/ProductLearningJobTotals"
        items:
          type: array
          items:
            $ref: "#/components/schemas/ProductLearningJobItem"
    KnowledgeCreateInput:
      type: object
      required: [shopId, question, answer]
      properties:
        shopId: { type: string }
        productId: { type: string }
        question: { type: string }
        answer: { type: string }
        effectiveFrom: { type: string, format: date-time }
        effectiveTo: { type: string, format: date-time }
    KnowledgeUpdateInput:
      type: object
      properties:
        shopId: { type: string }
        question: { type: string }
        answer: { type: string }
        effectiveFrom: { type: string, format: date-time }
        effectiveTo: { type: string, format: date-time }
        businessStatus:
          $ref: "#/components/schemas/KnowledgeBusinessStatus"
    KnowledgeBusinessStatus:
      type: string
      enum: [DRAFT, ENABLED, DISABLED, OUTDATED, CONFLICTED, DELETED]
    KnowledgeItem:
      type: object
      required: [id, shopId, scope, sourceType, businessStatus]
      properties:
        id: { type: string }
        shopId: { type: string }
        productId: { type: string }
        scope: { type: string, enum: [STORE, PRODUCT] }
        sourceType: { type: string, enum: [MANUAL, HUMAN_REVIEWED, AUTO_LEARNED] }
        businessStatus:
          $ref: "#/components/schemas/KnowledgeBusinessStatus"
        activeVersion:
          $ref: "#/components/schemas/KnowledgeVersion"
    KnowledgeVersion:
      type: object
      required: [id, version, question, answer, indexStatus]
      properties:
        id: { type: string }
        version: { type: integer }
        question: { type: string }
        answer: { type: string }
        indexStatus: { type: string, enum: [PENDING, INDEXING, READY, FAILED] }
        effectiveFrom: { type: string, format: date-time }
        effectiveTo: { type: string, format: date-time }
    KnowledgeImportJob:
      type: object
      required: [id, shopId, status, totals, rows]
      properties:
        id: { type: string }
        shopId: { type: string }
        status: { type: string }
        totals:
          type: object
          required: [total, valid, duplicate, conflict, error]
          properties:
            total: { type: integer }
            valid: { type: integer }
            duplicate: { type: integer }
            conflict: { type: integer }
            error: { type: integer }
        committedAt: { type: string, format: date-time, nullable: true }
        rows:
          type: array
          items:
            $ref: "#/components/schemas/KnowledgeImportRow"
    KnowledgeImportAccepted:
      type: object
      required: [operationId, importId, status]
      properties:
        operationId: { type: string }
        importId: { type: string }
        status: { type: string, enum: [ACCEPTED, QUEUED] }
    KnowledgeImportRow:
      type: object
      required: [rowNumber, scope, question, answer, status]
      properties:
        rowNumber: { type: integer }
        scope: { type: string, enum: [STORE, PRODUCT] }
        productId: { type: string, nullable: true }
        productExternalId: { type: string, nullable: true }
        question: { type: string }
        answer: { type: string }
        status: { type: string, enum: [VALID, DUPLICATE, CONFLICT, ERROR, COMMITTED, SKIPPED] }
        reason: { type: string, nullable: true }
        committedKnowledgeItemId: { type: string, nullable: true }
    KnowledgeSearchInput:
      type: object
      required: [shopId, query]
      properties:
        shopId: { type: string }
        query: { type: string }
        scope: { type: string, enum: [STORE, PRODUCT] }
        productId: { type: string }
        topK: { type: integer, minimum: 1, maximum: 3 }
    ReplyEvidenceSnapshot:
      type: object
      required: [itemId, versionId, version, source, scope, contentSnapshot, retrievalScore]
      properties:
        itemId: { type: string }
        versionId: { type: string }
        version: { type: integer }
        source: { type: string, enum: [MANUAL, HUMAN_REVIEWED, AUTO_LEARNED] }
        scope: { type: string, enum: [STORE, PRODUCT] }
        productId: { type: string, nullable: true }
        contentSnapshot:
          type: object
          required: [question, answer]
          properties:
            question: { type: string }
            answer: { type: string }
        retrievalScore: { type: number }
    KnowledgeSearchResult:
      type: object
      required: [status, evidence, conflictItemIds]
      properties:
        status: { type: string, enum: [EVIDENCE, NO_EVIDENCE, CONFLICTED, DYNAMIC_FACT_REQUIRED] }
        evidence:
          type: array
          items:
            $ref: "#/components/schemas/ReplyEvidenceSnapshot"
        conflictItemIds:
          type: array
          items: { type: string }
    KnowledgeRevisionResult:
      type: object
      required: [status, knowledgeId, version]
      properties:
        status: { type: string, enum: [UPDATED, CONFLICTED] }
        knowledgeId: { type: string }
        activeVersionId: { type: string, nullable: true }
        version:
          $ref: "#/components/schemas/KnowledgeVersion"
    KnowledgeConflictSide:
      type: object
      required: [itemId, versionId, question, answer]
      properties:
        itemId: { type: string }
        versionId: { type: string }
        version: { type: integer }
        question: { type: string }
        answer: { type: string }
        indexStatus: { type: string, enum: [PENDING, INDEXING, READY, FAILED] }
    KnowledgeCandidateStatus:
      type: string
      enum: [PENDING, APPROVED, PUBLISHED, REJECTED, DUPLICATE, CONFLICTED]
    KnowledgeConflictStatus:
      type: string
      enum: [OPEN, RESOLVED, IGNORED]
...
    KnowledgeCandidate:
      type: object
      required: [id, source, proposedQuestion, proposedAnswer, status]
      properties:
        id: { type: string }
        shopId: { type: string }
        source: { type: string }
        proposedQuestion: { type: string }
        proposedAnswer: { type: string }
        productId: { type: string, nullable: true }
        status:
          $ref: "#/components/schemas/KnowledgeCandidateStatus"
        duplicateOfId: { type: string, nullable: true }
        conflictWithId: { type: string, nullable: true }
        updatedAt: { type: string, format: date-time }
    KnowledgeConflict:
      type: object
      required: [id, leftVersionId, rightVersionId, status]
      properties:
        id: { type: string }
        shopId: { type: string }
        leftItemId: { type: string }
        rightItemId: { type: string }
        leftVersionId: { type: string }
        rightVersionId: { type: string }
        left:
          $ref: "#/components/schemas/KnowledgeConflictSide"
        right:
          $ref: "#/components/schemas/KnowledgeConflictSide"
        status:
          $ref: "#/components/schemas/KnowledgeConflictStatus"
        resolution: { type: object, nullable: true }
        resolvedAt: { type: string, format: date-time, nullable: true }
        updatedAt: { type: string, format: date-time }
...
    OperationAccepted:
      type: object
      required: [operationId, status]
      properties:
        operationId: { type: string }
        status: { type: string, enum: [ACCEPTED, QUEUED] }
...